backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
dj-database-url==2.1.0
python-dotenv==1.0.0
channels==4.0.0
channels-redis==4.1.0
twilio==8.12.0
pytest==7.4.0
pytest-django==4.5.2
drf-spectacular==0.26.3
python-json-logger==2.0.7
Pillow==10.1.0
requests==2.31.0
urllib3==2.0.7
cryptography==42.0.5
requests-oauthlib==1.3.1
pyotp==2.9.0
qrcode==7.4.2
django-templated-mail==1.1.1
zxcvbn==4.4.28
django-ratelimit==3.0.1
django-csp==4.0
fcm-django==2.0.0
firebase-admin==6.1.0
python-decouple==3.6
celery==5.2.7
redis==4.5.3
croniter==1.4.1
xlsxwriter==3.1.9
